replaced all the int datatypes with Long due to int cannot store the null value. completed the service implementation.
Will complete the controller in next commit

+++ todo/src/main/java/com/todoback/todo/repositories/TodoRepo.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.todoback.todo.model.Todo;
 
-public interface TodoRepo extends JpaRepository<Todo, Integer> {
+public interface TodoRepo extends JpaRepository<Todo, Long> {
 	List<Todo> findTodoByCategoryId(Long categoryId);
 }
